requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
pyyaml==6.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
aiofiles==23.2.1
slowapi==0.1.9
